added character select screen
Added new scene, which shows a deck of 6 cards. Press 's' to shuffle, and 'k' to present the top three cards. The shuffling animation is a lie, it does nothing.

The deck is a bunch of PlayerCharacters with the healthbars hidden. I couldn't figure out how to get a random CharacterObject asset into each PlayerCharacter; I need to sleep. Hopefully there's enough of a base there to work off of.